Fix image data error
I'm not 100% on this but I believe that the memory location the wasm
bungle was using was changing out on me and resulting in the App
throwing upon trying to use it in the canvas. I think I fixed it by
always re-building the image data before putting it into the canvas.
This may need to be optimized.