Refine wasm build & bridge process
Wasm-pack is doing a lot to facilitate packaging for npm libraries which
is not what I am interested it, so I am refining this process to use the
underlieing wasm-bindgen tool. Also I am putting the wasm access into a
web worker and adding this WasmBridge class to access that worker threw
a promise based architecture.