Make html/css totally static
I'm suddenly not sure if all of this react is actually necessary. The
only dynamic bits are the canvas/worker code and the controls and most
of that has been pulled out of react anyways so I'm going to strip all
of that out and see if it is worth it.