Initialize brain control
This is a personal plugin for my Obsidian knowledge base. The idea is to
add functionality to that tool without trusting random plugins.

I don't really have a lot of personal info in there, but I don't trust
myself not to put personal info in there so I feel better knowing that
the code belongs to me and Obsidian themselves.